requirements.txt
crewai
python-dotenv
crewai_tools
langchain_openai
langchain_community
langchain_huggingface
load_dotenv